libelf: use an unsigned type internally to represent ELF classes.
This should be an internal-only change.  External APIs that specify
or return ELF class information (such as gelf_getclass(3) and
gelf_newehdr(3)) continue to use a signed values, as before.


git-svn-id: svn://svn.code.sf.net/p/elftoolchain/code/trunk@4074 95820547-d848-0410-985e-9ae8fe0fa350

+++ libelf/gelf_getclass.c
@@ -37,5 +37,5 @@ ELFTC_VCSID("$Id$");
 int
 gelf_getclass(Elf *e)
 {
-	return (e != NULL ? e->e_class : ELFCLASSNONE);
+	return (int) (e != NULL ? e->e_class : ELFCLASSNONE);
 }
